finished initial interpreter and query function. works with js. Need to swift bindings, adding V + E, and handling field additions and anon traversals

+++ helixc/src/parser/helix_parser.rs
@@ -65,7 +65,7 @@ pub enum Expression {
     Traversal(Box<Traversal>),
     Identifier(String),
     StringLiteral(String),
-    NumberLiteral(i64),
+    NumberLiteral(i32),
     BooleanLiteral(bool),
     Exists(Box<Traversal>),
 }
